Move and rename modules
Making "System.Metrics.Static" (now "System.Metrics") the only exposed
interface.
